chore: Tweaked testing pages
